[#7] Add pagination util
